// frontend/src/api/schema.ts
export interface components {
    schemas: {
// ...
        CreateReportFromFormRequest: {
            patient: components["schemas"]["PatientFormData"];
            doctor: components["schemas"]["DoctorFormData"];
            medication: components["schemas"]["MedicationFormData"];
            adverse_effect: components["schemas"]["AdverseEffectFormData"];
            additional_info: components["schemas"]["FilesFormData"];
        };
// ...
        SeverityLevel: "mild" | "moderate" | "severe" | "life-threatening";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}